cleanup application store data
